Insert into index during chunk compression
If there is an index on the compressed chunk, insert into it while
inserting the heap data rather than reindexing the relation at the
end. This reduces the amount of locking on the compressed chunk
index which created issues when merging chunks and should help
with the future updates of compressed data.